1.1.2 Open and explore the starter scene
2. Open and explore the starter scene

Before you run the project, you should choose a more interesting environment to experience in VR.

1.  Rename and open the Starter Scene:

    From the Project window, open the Scenes folder.
    Rename the “Create-with-VR_Starter-Scene” as “[Your_Name] Room”.
    Double-click on your renamed scene to open it.

2.  Explore the contents of this scene and how it differs from a typical empty Scene:

    From the Hierarchy, select the XR Rig object and inspect the XR Origin component.
    Select XR Rig > Camera Offset > Main Camera and inspect the Tracked Pose Driver component.
    Select either the LeftHand Controller or RightHand Controller object and inspect the XR Controller component.
    Select the Input Action Manager object and inspect the Input Action Manager component.

In the Hierarchy window on the left, the Main Camera GameObject is a child object of the XR Rig.  In the Project window, the Scenes folder is open with “Carl Scene” highlighted.
Select image to expand

You should now have your new scene open. You should also have a broad understanding of the components of the scene that make it different from a typical Unity scene.

Note:  This scene and its settings have been pre-configured in this starter project, since they are required for VR development in this course. If you want to learn how to configure a VR-ready project similar to this from scratch, there are instructions in this tutorial: Create a VR Starter Project from Scratch.