1.1.3 Add a room and background
This empty scene would be boring to explore in VR, so you should add a room that will be a bit more interesting and will provide more perspective when you’re in VR.

1.  Add a room to the scene:

    In the Project window, expand Course Library > _Prefabs > Rooms.
    Drag one of the Room_[style] prefabs into the Hierarchy.
    From the Hierarchy, delete the Plane object.

2.  Add an environment outside the room’s windows:

    Open the Course Library > _Prefabs > Environments folder
    Drag one Foreground object and one Background object into the Hierarchy.

3.  Adjust the sunlight in the room:

    Change the X and Y rotation of the Directional Light object to change the way sunlight enters your room.

In the Hierarchy window on the left, 3 objects are highlighted: Room_Modern, Foreground_Buildings, and Background_Buildings. In the Scene view, there is a room floating in the air, with buildings in the background.
Select image to expand

You should now have room, foreground, and background objects in your scene from the course library, with sunlight entering the room at the desired angle.