1.1.5 Test in VR through Unity
If you are using a Windows computer, follow this step to test on your device. Otherwise, skip this step.
If you are using a Mac for development, this step is not relevant. You will learn how to test your app in the next step when you build your app directly onto your device.
In order to run the project on your device through Unity, you need to install the correct plug-in for development. OpenXR is an open source API that connects Unity to supported VR hardware devices. It is recommended that you use this plug-in, since it allows you to deploy to multiple devices through a single, standardized API.

1.  Install the OpenXR Plugin for desktop testing:

    From the top menu, select Edit > Project Settings, then select the XR Plug-in Management panel from the sidebar.
    In the Windows, Mac, Linux tab, select OpenXR from the list of available Plug-in Providers to install that plug-in.

Select image to expand

2.  Resolve any warnings by setting up an interaction profile.

    After adding the OpenXR plugin, there may be a warning or error icon that appears next to the plugin Name.
    If there are no warnings or errors, you can skip to step 3 below to connect your device through the Quest Link Software.
    If there is a warning, click on the warning icon to open the OpenXR Project Validation window, which will tell you that you need to add an interaction profile for the device you’re using.
    Select the Edit button to open the OpenXR settings panel.

Select image to expand

    In the Windows, Mac, Linux tab, make sure the Oculus Touch Controller Profile appears in the list of Interaction Profiles, then enable all available OpenXR Feature Groups.
    If you are using a different device, such as the Valve Index or HTC Vive, add its interaction profile instead.

Select image to expand

    There should no longer be any warnings in the XR Plugin Management panel. If there are, select them and follow the recommended steps to resolve them.
    If you are having trouble resolving warnings or errors, or if you need help troubleshooting, check out this detailed documentation on configuring the OpenXR plugin.

3.  Connect your device through the Quest Link Software:

    Make sure your device is plugged into your computer using a compatible cable.
    Make sure the Quest App is running and has successfully recognized your device.
    If you have not yet downloaded the Quest App, you find it on the Link Setup page.

Desktop Oculus App with Quest 2 and Touch shown as Connected.
Select image to expand

4.  Test the project on your device:

    Make sure the XR Device Simulator in your scene is disabled. If it is enabled, your head tracking will not work.
    Select Play in the Unity Editor and put on your headset.

Game View displaying two spherical hand models with the Play button highlighted at the top of the screen. The XR Device Simulator is highlighted in the Hierarchy, since it is deactivated and greyed out.
Select image to expand

On a Windows computer, you should now be able to quickly test your app on your device through Unity.