1.2.1 Fill the room with furniture
You now have a room, but it’s empty. You need to add spaces in the room where VR interactions can take place.

1.  Locate the furniture prefabs:

    In the Project window, open Course Library > _Prefabs.

2.  Define an entryway:

    In one corner of the room, add a mirror object and hook object.

3.  Define a living area:

    In one half of the room, add a seating object, a television object, and a fireplace object.

4.  Define another area:

    Add a rug object and either a table object or storage object to provide a surface.

Scene view displaying layout of furnished room.
Select image to expand

You should now have: an entry area with a hook and a mirror; a seating area, a living area with a television and a fireplace; and another area with some surface space to hold objects.